refactor: update import package path and fix the TokenizerChain as it used old constructors etc
